+++ frontend/src/constants/record.js
@@ -1,1 +1,2 @@
 export const SET_SIGNATURE_ASSETS = 'SET_SIGNATURE_ASSETS';
+export const CLEAR_SIGNATURE_ASSETS = 'CLEAR_SIGNATURE_ASSETS';
